changed from long int to long long unsigned int

+++ choose.cpp
@@ -1,8 +1,4 @@
 #include "permute.cpp"
-long int choose(long int n, long int k){
+unsigned long long int choose(long int n, long int k){
 	return permute(n,k)/permute(k,k);
 }
-
-
-
-
